requirements.txt
boto3>=1.26.0
botocore>=1.29.0

# FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Database and AWS
aioboto3>=12.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# WebSocket support
websockets>=12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
hypothesis>=6.88.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.6.0